+++ apps/extension/src/message/popup.ts
@@ -6,7 +6,7 @@ import type {
 } from '@penumbra-zone/types/src/internal-msg/shared';
 import type { UserChoice } from '@penumbra-zone/types/src/user-choice';
 import type { Jsonified } from '@penumbra-zone/types/src/jsonified';
-import { OriginRecord } from '@penumbra-zone/storage/src/chrome/local';
+import { OriginRecord } from '@penumbra-zone/types/src/local-storage';
 
 export enum PopupType {
   TxApproval = 'TxApproval',
