+++ apps/extension/src/state/network.test.ts
@@ -1,12 +1,12 @@
 import { beforeEach, describe, expect, test } from 'vitest';
 import { create, StoreApi, UseBoundStore } from 'zustand';
 import { AllSlices, initializeStore } from '.';
-import { LocalStorageState } from '@penumbra-zone/storage/src/chrome/local';
 import {
   mockLocalExtStorage,
   mockSessionExtStorage,
 } from '@penumbra-zone/storage/src/chrome/test-utils/mock';
 import { ExtensionStorage } from '@penumbra-zone/storage/src/chrome/base';
+import { LocalStorageState } from '@penumbra-zone/types/src/local-storage';
 
 describe('Network Slice', () => {
   let useStore: UseBoundStore<StoreApi<AllSlices>>;
