+++ apps/extension/src/state/password.ts
@@ -1,8 +1,8 @@
 import { AllSlices, SliceCreator } from '.';
 import { Key, KeyJson, KeyPrint } from '@penumbra-zone/crypto-web/src/encryption';
 import { ExtensionStorage } from '@penumbra-zone/storage/src/chrome/base';
-import { LocalStorageState } from '@penumbra-zone/storage/src/chrome/local';
 import { SessionStorageState } from '@penumbra-zone/storage/src/chrome/session';
+import { LocalStorageState } from '@penumbra-zone/types/src/local-storage';
 
 // Documentation in /docs/custody.md
