+++ apps/extension/src/state/persist.ts
@@ -2,10 +2,11 @@ import { StateCreator, StoreMutatorIdentifier } from 'zustand';
 import { AllSlices } from '.';
 import { produce } from 'immer';
 
-import { localExtStorage, LocalStorageState } from '@penumbra-zone/storage/src/chrome/local';
+import { localExtStorage } from '@penumbra-zone/storage/src/chrome/local';
 import { sessionExtStorage, SessionStorageState } from '@penumbra-zone/storage/src/chrome/session';
 import { StorageItem } from '@penumbra-zone/storage/src/chrome/base';
 import { walletsFromJson } from '@penumbra-zone/types/src/wallet';
+import { LocalStorageState } from '@penumbra-zone/types/src/local-storage';
 
 export type Middleware = <
   T,
