+++ apps/extension/src/state/wallets.ts
@@ -8,13 +8,13 @@ import {
 } from '@penumbra-zone/wasm/src/keys';
 import { Key } from '@penumbra-zone/crypto-web/src/encryption';
 import { ExtensionStorage } from '@penumbra-zone/storage/src/chrome/base';
-import { LocalStorageState } from '@penumbra-zone/storage/src/chrome/local';
 import {
   Address,
   FullViewingKey,
   SpendKey,
 } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/keys/v1/keys_pb';
 import { Wallet, WalletCreate } from '@penumbra-zone/types/src/wallet';
+import { LocalStorageState } from '@penumbra-zone/types/src/local-storage';
 
 export interface WalletsSlice {
   all: Wallet[];
