+++ packages/storage/src/chrome/test-utils/mock.ts
@@ -1,6 +1,7 @@
 import { ExtensionStorage, IStorage } from '../base';
-import { localDefaults, LocalStorageState } from '../local';
+import { localDefaults } from '../local';
 import { sessionDefaults, SessionStorageState } from '../session';
+import { LocalStorageState } from '@penumbra-zone/types/src/local-storage';
 
 // Helpful for testing interactions with session & local storage
 export class MockStorageArea implements IStorage {
